fix: audit-driven correctness fixes across kernel, ops, infra, state
Findings from an 8-agent parallel audit. Fixes are limited to confirmed
correctness bugs; design-level / borderline findings are listed in the
local audit checklist for follow-up.

Kernel + dispatch
* _kernel/registry/primitive: do not clear _PENDING_COMPOSES before
  raising; retries would otherwise re-enter with an empty pending dict
  and "succeed" with a half-resolved registry.
* _kernel/extension/version.is_compatible: use .get(domain, ()) so the
  documented non-raising contract holds for unknown domains.
* _kernel/lifecycle/invoke + run: capture worker stderr and surface its
  tail in the error envelope when a worker fails to return a parseable
  report — previously discarded.
* cli/spawn: mark the ``run`` envelope idempotent=False (it spawns a
  fresh-context worker; the default True misrepresented the contract).
* cli/_dispatch._emit_dry_run: check isinstance(dict) before hasattr,
  so passthrough keys that collide with dict method names (items/keys/
  get) don't resolve to bound methods.

Infra
* infra/parsing.parse_mem_to_gb: re.fullmatch so trailing garbage no
  longer parses as a numeric prefix (``5days`` -> 5MB previously).
* infra/backends/sge: add the comma-in-value guard for ``qsub -v`` that
  the SLURM backend already enforces and the SLURM comment references
  by name. Forwarded env values containing commas now fail fast.
* infra/backends/query: refuse ``qstat -u ""`` when no user identity is
  available — some SGE versions interpret empty user as "no filter".
* infra/remote._scp (deploy_runtime): pass ``-o BatchMode=yes`` so a
  missing key fails fast instead of blocking on a password prompt
  (matches _scp_pull and ssh_run).

State + ops
* state/run_record: HPC_JOURNAL_DIR="" now falls back to default
  cleanly; previously the empty-string env value silently bypassed
  the v3 "env wins" precedence at both import-time and runtime.
* ops/monitor/summary: read the delta count from nc[0]/nf[0], not
  len(nc)/len(nf). monitor_flow writes ``[cur - prv]`` (length-1 list
  with the delta as its value); the consumer was counting the list
  length, so any progress tick rendered ``+1 complete`` regardless of
  the actual delta.

Incorporation + models
* incorporation/build/template: fix scaffold_dir path —
  _PACKAGE_ROOT/template/scaffold did not exist; the real location
  is _PACKAGE_ROOT/incorporation/template/scaffold. Every
  ``build-template`` call previously raised FileNotFoundError.
* incorporation/template/signature: kwonly default extraction now
  uses ``dflt is not None`` instead of a truthy check, so falsy
  literal defaults (0, 0.0, "", False, []) no longer collapse to
  None and skew run_signature_sha.
* models/mapreduce/reduce/metrics: sum cpu_s / gpu_s / elapsed_s as
  floats instead of int-truncating per task, so sub-second tasks no
  longer round down before aggregation.

Scripts
* scripts/lint_skill_command_sync: replace ``hpc-agent run\b`` with
  ``hpc-agent run(?![\w-])`` — \b matched between ``run`` and ``-``
  so typos like ``hpc-agent run-time`` falsely satisfied the routing
  requirement.
* hpc-agent-pro/scripts/train_wait_predictor: coerce naive
  submit/start/floor datetimes to UTC-aware before mixing with the
  tz-aware snapshot timestamps, which were raising TypeError.

Tests updated to match producer-aligned consumer behaviour (summary
delta encoding; scp BatchMode flag layout).

+++ src/hpc_agent/models/mapreduce/reduce/metrics.py
@@ -270,16 +270,20 @@ def reduce_resource_usage(tasks: dict[str, dict] | dict[int, dict]) -> dict:
             "tasks_counted": int, # number of tasks that contributed nonzero elapsed_s
         }
     """
-    total_cpu_s = 0
-    total_gpu_s = 0
-    total_elapsed_s = 0
+    total_cpu_s = 0.0
+    total_gpu_s = 0.0
+    total_elapsed_s = 0.0
     counted = 0
     for info in (tasks or {}).values():
         if not isinstance(info, dict):
             continue
-        elapsed = int(info.get("elapsed_s", 0) or 0)
-        cpu = int(info.get("cpu_s", 0) or 0)
-        gpu = int(info.get("gpu_s", 0) or 0)
+        # Per-task resource values can be fractional (sub-second tasks,
+        # GPU-seconds during partial allocation windows). Coercing to
+        # int at the task level truncates the fraction before summing
+        # and silently under-counts. Sum as float; round at the end.
+        elapsed = float(info.get("elapsed_s", 0) or 0)
+        cpu = float(info.get("cpu_s", 0) or 0)
+        gpu = float(info.get("gpu_s", 0) or 0)
         total_elapsed_s += elapsed
         total_cpu_s += cpu
         total_gpu_s += gpu
